Filter suites for each manager

+++ src/Controller/Backend/SuiteController.php
@@ -8,6 +8,7 @@ use App\Repository\SuiteRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\UnauthorizedHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 
 #[Route('/admin/suites')]
@@ -17,7 +18,7 @@ class SuiteController extends AbstractController
     public function index(SuiteRepository $suiteRepository): Response
     {
         return $this->render('backend/suite/index.html.twig', [
-            'suites' => $suiteRepository->findAll(),
+            'suites' => $suiteRepository->findAllForManager($this->getUser()),
         ]);
     }
 
@@ -42,6 +43,10 @@ class SuiteController extends AbstractController
     #[Route('/{id}', name: 'backend_suite_show', methods: ['GET'])]
     public function show(Suite $suite): Response
     {
+        if ($suite->getHotel()->getOwner() != $this->getUser()) {
+            return $this->redirectToRoute('backend_suite_index', [], Response::HTTP_FORBIDDEN);
+        }
+
         return $this->render('backend/suite/show.html.twig', [
             'suite' => $suite,
         ]);
@@ -50,6 +55,10 @@ class SuiteController extends AbstractController
     #[Route('/{id}/edition', name: 'backend_suite_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Suite $suite, SuiteRepository $suiteRepository): Response
     {
+        if ($suite->getHotel()->getOwner() != $this->getUser()) {
+            return $this->redirectToRoute('backend_suite_index', [], Response::HTTP_FORBIDDEN);
+        }
+
         $form = $this->createForm(SuiteType::class, $suite);
         $form->handleRequest($request);
 
@@ -67,6 +76,10 @@ class SuiteController extends AbstractController
     #[Route('/{id}', name: 'backend_suite_delete', methods: ['POST'])]
     public function delete(Request $request, Suite $suite, SuiteRepository $suiteRepository): Response
     {
+        if ($suite->getHotel()->getOwner() != $this->getUser()) {
+            return $this->redirectToRoute('backend_suite_index', [], Response::HTTP_FORBIDDEN);
+        }
+
         if ($this->isCsrfTokenValid('delete'.$suite->getId(), $request->request->get('_token'))) {
             $suiteRepository->remove($suite);
         }
